+++ src/main/java/online/kingdomkeys/kingdomkeys/client/render/entity/ElementalMusicalHeartlessRenderer.java
@@ -14,8 +14,6 @@ import online.kingdomkeys.kingdomkeys.entity.mob.EmeraldBluesEntity;
 
 public class ElementalMusicalHeartlessRenderer extends MobRenderer<BaseElementalMusicalHeartlessEntity, ElementalMusicalHeartlessModel<BaseElementalMusicalHeartlessEntity>> { //my god that's a long one
 
-    static final double MAX = 200;
-
     public ElementalMusicalHeartlessRenderer(EntityRendererProvider.Context context) {
         super(context, new ElementalMusicalHeartlessModel<>(context.bakeLayer(ElementalMusicalHeartlessModel.LAYER_LOCATION)), 0.35F);
         this.addLayer(new HeartlessEyesLayerRenderer<>(this, KingdomKeys.rl("textures/entity/mob/musical_heartless_eyes.png")));
@@ -29,12 +27,6 @@ public class ElementalMusicalHeartlessRenderer extends MobRenderer<BaseElemental
     @Override
     protected void scale(BaseElementalMusicalHeartlessEntity entitylivingbaseIn, PoseStack matrixStackIn, float partialTickTime) {
         matrixStackIn.scale(1F, 1F, 1F);
-        double pos = entitylivingbaseIn.tickCount % MAX / (MAX /2D);
-        if(entitylivingbaseIn.tickCount % MAX < (MAX / 2)) {
-        	matrixStackIn.translate(0, pos*0.6, 0);
-        } else {
-        	matrixStackIn.translate(0, (MAX - entitylivingbaseIn.tickCount % MAX) / (MAX / 2D) * 0.6, 0);
-        }
         if(entitylivingbaseIn instanceof EmeraldBluesEntity) {
             matrixStackIn.scale(1.1F, 1.1F, 1.1F);
         }