+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/LanguageENUS.java
@@ -1934,6 +1934,10 @@ public class LanguageENUS extends KKLanguageProvider {
         addItem(ModEntities.SOLDIER_EGG, "Soldier Spawn Egg");
         addItem(ModEntities.DESERTER_EGG, "Deserter Spawn Egg");
         addItem(ModEntities.COMMANDER_EGG, "Commander Spawn Egg");
+        addItem(ModEntities.AIR_SOLDIER_EGG, "Air Soldier Spawn Egg");
+        addItem(ModEntities.DEFENDER_EGG, "Defender Spawn Egg");
+        addItem(ModEntities.NEOSHADOW_EGG, "Neoshadow Spawn Egg");
+        addItem(ModEntities.NOVASHADOW_EGG, "Novashadow Spawn Egg");
         addItem(ModEntities.WHITE_MUSHROOM_EGG, "White Mushroom Spawn Egg");
         addItem(ModEntities.BLACK_FUNGUS_EGG, "Black Fungus Spawn Egg");
         addItem(ModEntities.BLOX_BUG_EGG, "Blox Bug Spawn Egg");
@@ -2278,6 +2282,10 @@ public class LanguageENUS extends KKLanguageProvider {
         addEntityType(ModEntities.TYPE_SOLDIER, "Soldier");
         addEntityType(ModEntities.TYPE_DESERTER, "Deserter");
         addEntityType(ModEntities.TYPE_COMMANDER, "Commander");
+        addEntityType(ModEntities.TYPE_AIR_SOLDIER, "Air Soldier");
+        addEntityType(ModEntities.TYPE_DEFENDER, "Defender");
+        addEntityType(ModEntities.TYPE_NEOSHADOW, "Neoshadow");
+        addEntityType(ModEntities.TYPE_NOVASHADOW, "Novashadow");
         addEntityType(ModEntities.TYPE_WHITE_MUSHROOM, "White Mushroom");
         addEntityType(ModEntities.TYPE_BLACK_FUNGUS, "Black Fungus");
         addEntityType(ModEntities.TYPE_BLOX_BUG, "Blox Bug");