+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Added:
 * Realm of Darkness arches.
+* World Map dimension.
+* Air Soldier, Defender, Neoshadow and Novashadow
 
 ## Changed:
 * Lowered meteorite frequency.